llm_orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
python-dotenv==1.0.0
httpx==0.25.2
sniffio>=1.3.0 

